+++ app/src/main/java/nextstep/jwp/controller/Controller.java
@@ -2,8 +2,8 @@ package nextstep.jwp.controller;
 
 import java.io.FileNotFoundException;
 
-import nextstep.jwp.handler.HttpRequest;
-import nextstep.jwp.handler.HttpResponse;
+import nextstep.jwp.handler.request.HttpRequest;
+import nextstep.jwp.handler.response.HttpResponse;
 
 public interface Controller {
