+++ app/src/main/java/nextstep/jwp/handler/constant/HttpStatus.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.handler;
+package nextstep.jwp.handler.constant;
 
 public enum HttpStatus {
