+++ app/src/main/java/nextstep/jwp/handler/response/HttpResponse.java
@@ -1,7 +1,12 @@
-package nextstep.jwp.handler;
+package nextstep.jwp.handler.response;
 
 import java.util.HashMap;
 
+import nextstep.jwp.handler.HttpBody;
+import nextstep.jwp.handler.HttpHeader;
+import nextstep.jwp.handler.constant.HttpStatus;
+import nextstep.jwp.handler.request.HttpRequest;
+import nextstep.jwp.handler.response.ResponseLine;
 import nextstep.jwp.util.ContentType;
 import nextstep.jwp.util.File;
 
@@ -11,7 +16,7 @@ public class HttpResponse {
     private final ResponseLine responseLine;
     private final HttpHeader header;
     private HttpBody body;
-    
+
     public HttpResponse(HttpRequest httpRequest) {
         this.responseLine = new ResponseLine(httpRequest.getHttpVersion());
         this.header = new HttpHeader(new HashMap<>());
